EDIT search to restore autocomplete for filter.
DSpace has autocomplete for filter terms. This feature was lost in the
new search but has now been restored. When entering custom filters using
"Advanced Search", there should now be an autocomplete menu available
depending on what you type.

The javascript part that does most of the heavy lifting is just copied
and pasted from the old search. It's using jQuery UI's autocomplete
function.

This is task 13 of task list #62.