EDIT search to persist list/tile view selection.
When you switch to the tile view, going to the next page makes the
search revert back to list view. Unfortunately, persisting this choice
means that toggling between views will now trigger a page refresh. As
all persistant information in search is being handled by form fields, we
need to trigger a form submit in order to store the fact that we changed
view types. This would be something I'd like to change architecturally
in the future, but don't have time right now.

This is task 14 in task list #62.