EDIT search result filter show subject hierarchy
In the results filter boxes, subjects are shown in their entirety, so if
we have a 3 level deep subject, we'd see "level 1 >>> level 2 >>> level
3". This is a lot of text, so it'd be nice if, for lower level subjects,
we can avoid having to put their parent subject. This requires
perserving the hierarchy of the subjects when displaying it.

The SubjectsJson class was renamed to the SubjectsTreeParser. This is
because I've expanded it to also parse the subjects for search. I really
couldn't figure out a good way to represent the subjects tree using the
standard data structures, so I've created SubjectEntry to help with
presenting the subjects tree.

The filter result boxes now uses + and - icons to indicate that the box
can be collapsed/shown. This is for consistency with the + and - icons
used on the search result "Details" button.

One thing that I didn't get to is allow the higher level subjects to
toggle show/hide the lower level subjects. Just realized that was part
of my initial image on how this worked, but had forgotten it while
writing the code.

This is for task 11 of task list #62.