EDIT search result filter collapsible subject tree
The subjects search result filter hierarchy is going to start getting
too long once the subjects start getting filled up. It would be nice if
the subtopics can be collapsed into the main topic. Unfortunately, to do
this required some terrible mess in rendering the subjects tree. So I've
split the subjects rendering out into its own component JSP.

I'm not sure how best to present the collapsible UI to the user.
Currently using a small button with the chevron to make it more obvious
to the user that it's a control.

Also switched the +/- icon for collapsible content into the up and down
chevron standards. Hoping to avoid confusion since the plus sign is
usually used as an "Add" operation.

This also marks a divergence with BioSpace, since BioSpace isn't using a
subject tree, it doesn't need this feature. I'm not sure what to do yet
for merging this feature into BioSpace.

This is for task 11 in task list 62.