EDIT searchbox to top nav bar, about page
Restore the searchbox that was in the top nav bar. Removed the searchbox
that was on the home page. This should allow faster access to search.

About page changed to prevent confusion with the main page. Also added a
breadcrumb on top of page.

This is task 100 in task list #61.