EDIT html5 validate on login, user registration.
If login fails, the user email gets cleared, causing users to have to
re-enter it. Preserve the data to save users some work.

When filling in the user registration field, making a mistake causes
lost of data. Try to prevent this by using html5 validation on the page
so that the form isn't submitted if there's an error.

Note, if required is passed in to the password field jsp component,
it'll add required to the password field. This unfortunately doesn't
play well with password autofill as the error message gets blocked by
the password autofill tooltip.

Partial task 3, 4 in tasks list #61. Partial because this probably needs
more work to completely implement in StatSpace.