EDIT MyDSpace, replace Open button with Edit&View
Save a click where instead of having to click Open and then Edit/View,
you can click Edit/View directly on the MyDSpace page.

Also added storing and showing the date & time that the submission was
started. This is intended for providng a way to sort the table later on.

Also removed the "Submitted By" field, as all submissions are by
you.

Note that I've discovered there's a "Supervisor" capability where you
can see submissions from the other users. This is what the "Submitted
By" field is intended for. I've made the supervisor submission table
separate from your own submissions, so that "Submitted By" still shows
up in the supervisor table.

This is task 13 in task list #61. Not counting this as a completion
until it works in StatSpace & BioSpace.