FIX Advanced Search errors out on form submit
If the "Advanced Search" is open and the user does any action that
requires a submit of the search form, it will error out if the user
hasn't filled in the filter fields. This is because DSpace tries to
apply a filter but finds that there isn't enough information to do so.

Workaround for now is to just hide the Advanced Search forms before
submit if the filters aren't filled in.