FIX repeatable fields are not saved in order.
The order that items are entered in isn't preserved on save. So if you
entered A, B, C for Pre-reqs, it might load up as C, A, B when you come
back to it. This has something to do with the weird sequence numbering
that the old edit-metadata.jsp form generates where the inputs are
presented in descending sequence.

Instead of each repeated element having a uniquely numbered name (a huge
pain to keep track of), we can just rely on ServletRequest converting
inputs that have the same name into arrays. So I've modified the API and
edit-metadata.jsp to use this method. This allows removing a lot of now
uneeded complicated code from edit-metadata.

Interestingly, it turns out the Remove button doesn't need it's own ID,
as it operates on its parent. Not sure why it had a unique ID in the
first place now.

Note, the same complicated numbering scheme still exists in search, so
we're not completely rid of this yet.

This implments task 40 in task list #64.