ADD custom metadata registry file ubc.xml
The UBC registry is for metadata that doesn't fit in either the original
dc or dcterms metadata that comes with DSpace. The initial ubc.xml only
specifies one field, ubc.resourceTypeOther, this is similar to
dc.subject.other but for resource types in that instead of being an
option selected from a dropdown, it's a textbox that the user enters
data into.

The more important addition brought about by this commit is the
StartupJob. Besides replaceing the original UpdateSolrIndexJob, it'll
update the registry database when there are changes in ubc.xml or other
registry files. This automates an operation originally meant to be run
from the command line DSpace tools

This is part of task 38 in #64.