ADD resource type "Other" custom entry field.
Selecting "Other" for resource type will prompt the user to enter a
custom resource type. The javascript for a custom subject has the same
functionality and was reused for this. The custom resource type is also
displayed in the same way as a custom subject: 'Other (custom entry
here)'.

Warning, just realized that both this custom resource type and custom
subject isn't indexed by the search engine as part of resource type or
subject since they're stored in different metadata fields.

Part of task 38 in task list #64.