EDIT profile form validation & instructor access
Add HTML5 form validation to the profile form to try to prevent data
loss on form submits.

The instructor access section was moved inline with the rest of the form
and made more obvious that it's an optional section which belongs with
the instructor information fields.

This completes task 3 & 28 of task list #61.