EDIT rm logged in search menu, add MyDSpace icon.
Logged in users have a search menu which is now unnecessary due to the
search box being restored.

About now shows up in the menu when logged in also.

Add an icon for MyDSpace in the menu to make it match with About.

Completes task 17 on task list #65. Also some tasks mentioned in the
document in #61, but not defined as a task.