EDIT search hide filters that are already in use
In "Filter Results", filters that have already been applied still show
up as options, leading to cases you can apply the same filter multiple
times.

Ideally, I'd like the applied filters to be marked differently somehow,
and not have to use the current applied filter listing method. But that
would be a pretty big project, so setting it aside for now and hopefully
can get around to it later.

This commit just hides the already applied filters so that they're
available as an option in the "Filter Results" listing.