EDIT search move search controls to sidebar
The big search box and the search filters section on the top is now
placed as part of the left sidebar. Bio hopes this makes the search
results to be more prominent. As part of that, the thumbnails were made
larger too.

The same subject listing that's on the home page is now also listed at
the bottom of the left sidebar. This required making the subject listing
code shared, so that code has been moved into SubjectInfo.

Interestingly, while doing this, I encountered a couple of issues with
the search filters which I had to first resolve in StatSpace and then
merged back in.  Notably, the search filter numbering broken when I
moved the search filter code and so I just got rid of the numbering
entirely. Also, another major issue is that filters which are currently
applied are still listed in the filters list. This makes it too easy for
the same filter to be applied multiple times. Quick work around was to
just hide them already applied filters, but I'd like to see a more user
friendly solution where the filters are marked as applied and we don't
have to use the current filter listing mechanism.

This completes task 44 in task list #66.