ADD contact us page, admin editable blurb system
The original request was for a curator editable Contact Us page. There's
already an editable "news" system in DSpace, it's meant for the original
DSpace home page. Since the news system wasn't being used, it was
co-opted for this purpose. Unfortunately, this means that you need to be
an administrator to make edits, but that shouldn't be too much of an
issue.

The edit news system was systematically renamed to "Edit Blurbs".

The home page blurb was also converted to be editable. Thus kinda
returning to the original editable DSpace homepage.

Note that the editable blurbs are actually just plain html files. The
only validation is on the file names, which are defined in
core-services.xml. I've converted the plain text area for editing into
a TinyMCE editor for user friendliness.

This completes task 41 of task list #66.