FIX home page, fade placeholder thumbnail
The featured articles section doesn't fade the placeholder thumbnail
like on the search pages.

Also increased the UBC xl logo canvas size so that it appears smaller
relative to other thumbnails, making it less noticable as befitting a
placeholder thumbnail.

Changed the news-top template so that it's centering works with tinymce.