ADD related materials section link other resources
The related materials section can now be configured to link to other
resources that the submitter has submitted.

This is an initial implementation and probably has a lot of bugs.

We're calling these links to other resources "related resources" to
distinguish them from other entries in related materials.

Related resources are distinguished by a unique header, the header is a
constant defined in ItemRetriever. The header is followed by the item
ID.

This is task 74 from task list #58.