FIX date & time displayed in UTC in many places
Date & time is stored in UTC by default. ItemRetriever converts the
display into server timezone. We should display it in whatever the
client's timezone is. The dspace:date tag was modifed for this purpose.

Since the dspace:date tag takes a DCDate, we needed to modify
ItemRetriever to return a DCDate. This produces an issue where we can't
use JSTL's empty check to see if the date is valid or not since I would
rather not modify DCDate in case it breaks something. So ItemRetriever
now returns null if the date is empty. This means that JSPs need to
check if the date exists before trying to use it, or we'll get an
Internal Server Error due to the null pointer.

Note that this was only fixed in areas that I've touched. Places still
running the default dspace code are not fixed.

Implements task 107 in #67.