FIX css print view, curator approve/reject border
Testers complain that the print view is all messed up.  Looks like this
is primarily due to Bootstrap printing out the links by default in the
print view. Used an override to disable this feature.

The curator approve/reject control section has messed up borders that
overlaps with the help text. It was using bootstrap's input groups
feature in a weird way. Replaced the entire section with a table so that
it'll maintain the proper borders.

Implements task 105 in #67.