FIX edit-metadata "Add More" insertion order
"Add More" was inserting elements after the first field instead of the
last field. This was due another change I made to fix duplicate IDs due
to cloning the field where I just removed the ID entirely.

Unfortunately, this breaks the insertion order as the insert selector
depends on the ID. So as a stupid but quick fix, I'm just now changing
the ID so that it'll append a 1 to the new ID.

Note: Need to have unique IDs or TinyMCE will break.

Fixes task 103 in #67.