EDIT edit-metadata related material tinymce editor
Related material's "Other" option used to be just a simple input text
box, it is now a TinyMCE enabled editor. This is mainly so that users
can put in URLs faster.

A lot of refactoring was done with this, namely splitting up the related
material options into their own JSPs to be included. This was something
I didn't have time for earlier but is necessary in order to reduce code
duplication.

Implements task 101 in #67.