ADD bidirectional resource linking
During the submission process, users can now link to items they've
submitted and check a box that can forces the other item to also link
back.

Note that for these links to other item/resource in DSpace, I'm calling
them related resources. Related materials refer to anything in the
dcterms.relation field, related resources refer to related materials
that are item links.

A new class was created to handle the complicated cases involved in
trying to keep the bidirectional related resource data consistent.
RelatedResourceManager has to be located in the API because there's
cleanup that has to be done in the Item delete. E.g.: deleting an item
would require us to clean up bidirectional links so we don't have
invalid references to non-existent items. Unfortunately, I ran out of
time and wasn't able to completely cover all such edge cases. So I've
made the code resistent to non-existent items instead. One such case is
if admin/curator deletes an archived item, the cleanup code wasn't built
with such a case in mind. The cleanup code was built with in-progress
submissions in mind, not already archived items.

Since regular users have no power to edit items, I had to disable the
authorization system when modifying the target item for bidirectional
links.

Note that a bidirectional link is stored separately in the target item
in the dc.relation.isreferencedby field. This is to easily tell if the
link was put in place by the submitter originally or if it was added at
a later date.

Implements task 74 in #67.