ADD How to Submit editable blurb page to top nav
Removed the news blurb html files, found out they were wiping the old
existing blurbs with the generic ones on deploy. This wipes out all
customizations that isn't committed in git, which makes the editable
blurb system pointless. The whole point was to make it easier for
non-coders to modify the content. Hopefully removing the blurb templates
from the build will stop this from happening.

The "How to Submit" page is just a regular blurb page. The addition of
this page on the top nav causes the search bar & user menu to be pushed
down to a second row even on the large view, so I shrunk the "My
BioSpace Submissions/Resources" back down to "My BioSpace", this at
least makes the top nav display in only 1 row on the large view.

Some minor refactoring, never did like the "ContactPage" name.

Implements task 46 in #68.