EDIT various small tweaks & changes.
Smaller tasks that I didn't want to split into small commits.

49. "My Accepted Submissions" page, rename "author" to "contributor",
"issue date" to "date created".

50. Change wording on acceptance of submission: "your submission has
been archived", replace archived with approved.

51. Show "Date Created" and "Date Approved". Perhaps with a tool tip on
"Date Approved" explaining that it's the date that the resource was
accepted into BioSpace.

52. Make the "Request Instructor Access" blurb editable. It still
mentions StatSpace.

54. Label for "Optional File Notes" on the upload page, remove
example and add a note saying that this is visible to users.

55. Submissions "Describe" step should break up the instructions to
be shorter to facilitate readability.

56. Add font size control to TinyMCE's big text editor.

57. Listing for subjects should be the same as keywords, with each
subject in its own cell.

58. Allow any registered user to make submissions, curator will
manually check to make sure user should be submitting.

59. Change sort option labels.
    "Sort by Title" should read "Sort by Title (alphabetical)"
    "Sort by Author" should read "Sort by Contributor (last name)"

60. Add instruction on file upload page to ask the user to upload an
image for thumbnail purposes. Indicate that the first thumbnailable file
uploaded will be used as the thumbnail.

Implements task 49,50,51,52,54,55,56,57,59,60 in #68.