FIX blurbs are lost on docker build/deploy
Blurbs uses DSpace news system where news are stored in individual
files. These files are not persisted across docker builds, which means
when we deploy a new version, they get overwritten. To fix this, we have
to store the blurbs in the database.

The blurbs are stored in "ubc.blurbs" metadata field in the first
community we find. This is not ideal, as it can't deal properly with
multiple communities if we ever start using that feature. But it's the
best way to implement it fast.

Blurbs are stored in a single metadata entry as a JSON object, indexed
by the type of blurb it is. Use BlurbManager for saving/retriving
blurbs.