EDIT various labels & messages
113. Edit your profile page: 'Can we change some grammatical errors on
the “Edit Your Profile” page?  "We recognize that there are individuals
in other roles that who make valuable contributions to statistics
education." Also, in the note next to the “Supervisor Contact” field,
“phone” and “email” should be in lower-case letters.'

114. Instructions on uploading videos: "If you are submitting a video as
part of your resource, you will have the option to submit an external
link, and/or to upload a file.  Supported video file formats include
[list of accepted file formats].  If you choose to upload a file, users
will be able to stream the video in StatSpace, and also to download it."

115. Resource Type help text currently says: "Select the resource
type(s) of the main content of the item (hold Ctrl/Cmd while clicking to
select multiple subjects".  Change "multiple subjects" to "multiple
resource types"

116. Bug: Error messages about missing metadata.* messages for new
metadata fields that we've added. This is used for metadata labels.  Not
sure if anything is using those labels, so might be safe to ignore.

Implements task 113,114,115,116 in task list #67.