EDIT s/material/resource/g, minor fixes
Consistently refer to resources in StatSpace as 'resource'. On the home
page, there is a tendency to refer to them as 'materials' instead of
'resources'.

It makes more sense to relabel "Access Restriction" as "Access" because
"Access Restriction: Everyone" can be mistaken as preventing everyone
from accessing the resource.

Also fixed URL generation for the comment pages so that it'll work on my
localhost dev instance.