FIX removing featured articles sometimes fails
This only seems like a sporadic failure because it depends on where you
click. If you click on the trash icon directly, it'll fail. But if you
click on the outer edges of the button, it'll succeed.

The bug doesn't seem to occur in Firefox, seems to only happen in Webkit
lineage browsers, such as Chrome.

It has to do with the event object passed to the callback, the event
object passed in Chrome turns out to be <span> that holds the trash
icon, hence the failure as the code expects it to be the <button>
element. Looks like I should've used "this" instead of "event.target" to
accurately get the element that fired off the event.