FIX grant/deny instructor permission button
If you click on the icon of the checkmark or the x, the button click
event will fail as it gets the wrong button id. Can't use the event
target to get the button id, as on webkit browsers, it gives you the
span holding the icon instead of the button. Should use the "this"
variable.