FIX prod pulling in latest instead of tagged ver
Turns out because the tagged version doesn't have the updated
.gitlab-ci.yml file, the pinning doesn't work. Used a temporary
workaround where we checked out the tagged version in branch
statspace_oldprod and then committed the updated gitlab config. Tagging
the statspace_oldprod branch then triggered a build that pulled in the
right container for prod.

The workaround statspace_oldprod branch can be deleted after we've got a
new version deployed to prod.