UPDATE openjdk7 to 8, tomcat8.0 to tomcat8.5
Tomcat 8.0 has been EOL since last year 2018, so we need to update to
latest 8.5. There's also a critical security advisory that won't allow
us to use Tomcat versions <=8.5.40.

The openjdk:7-jdk image uses Debian 8, both OpenJDK 7 and Debian 8 will
be EOL next year 2020. So I'm experimenting with updating to
openjdk:8-jdk, which uses Debian 9, and should be good to 2022.

The ghostscript upgrade hack is no longer necessary, as Debian 9 has
newer ghostscript packages that addresses the Microsoft PDF
incompatibility bug. The hack originally used Debian 9 packages anyway.

Also pulled in a fix for the geolocation database not being built in the
image that came from statspace. Not familiar with the fix, so have only
pulled in the Dockerfile part of the fix.