EDIT change metadata fields in simple item view
Customize the metadata fields shown in simple item view.

While the list of metadata fields is configured in dspace.cfg, for
deployment, we're going to have to update the chart's version of
dspace.cfg too.

Added a new field style html for fields like description which requires
html tags to be displayed. This involves modifying ItemTag.

Just noticed that the accessbility fields are under the namespace
'schema', they too should be moved to the 'ubc' namespace.