ADD blurb system which replaces the news system
The news system allows admins to edit certain parts of the website
without having to modify the JSPs. Unfortunately, this system stores the
edited content as text files, which are NOT persistent across kubernetes
deployments.

The blurb system replaces the news system by storing the content in
metadata. We create a dummy user and then store the content as part of
the dummy user's metadata. The "ubc.blurbs" metadata field was added for
this purpose.