EDIT select collection step also select community
The select collection step lists each collection as "COMMUNITY_NAME >
COLLECTION NAME". We want to change this so that the user selects the
community first from a drop down, then select a collection. This is
because each community has a lot of collections, so it can be hard to
locate in the original implementation.