FIX slow select collection step loading
When you click create submission, it loads the select collection step
first. The loading time has increased a lot with the introduction of the
two step community/collection selection. I think this has to be with the
use of TreeSet which gives us an alphabetically sorted Community list.
Even a get with TreeSet costs o(n). Replacing it with a LinkedHashSet to
see if it makes a difference.