EDIT basic style overhaul look more like ubc sites
Basic style overhaul to look more like the standard UBC theme.

The CWSEI logo has a static white background, so I have to use a white
background for the banner. There was a UBC crest next to it, but it
seems to make the banner a bit busy, so I've left it off for now.

The footer is very barebones cause I couldn't find an address or general
email address for CWSEI.

Moved basscss from CDN to local, as I seem to occasionally encounter
delays loading the library.