EDIT collection home show/sort Item Type & Subject
Collection home now shows an item's Item Type and Subject. The list of
items is sorted by Item Type by default.

There is a bug in the BrowseListTag where all field names are converted
to lower case. So metadata fields with camelCase names are all NOT
usable. Which is, shall we say, odd, because there's a lot of dcterm
fields using camelCase. So out of the box, we can't use any dcterm
fields in browse list?!  I had to make it not convert everything to
lower case to get ubc.itemType to work.

Unfortunately, the lower case conversion also applies to field labels,
that's why browse.menu.itemtype doesn't use itemType like it should. I
didn't want to touch BrowseListTag too much, so didn't try to get rid of
all lower case conversions just in case it breaks something else.