FIX slow select collection step loading
When there's a lot of collections, the select collection step takes a
long time to load. I initially thought this was due to my new code, but
on actually timing the code, found out it wasn't. It turns out that the
API for retrieving the collections list takes about 30s to complete.
Following the execution path, I found out that there's an option to
enable an "optimized" version of getting the collections list. Enabling
it cut the time down to 1s.

There's some caveats in the comments around this option, but I didn't
see any problems in my tests, so I've enabled it.