FIX wrong error when no file uploaded
Since we require at least 1 file to be uploaded, there should be a
proper error message if no files were uploaded and the user presses
next. Unfortunately, due to some really weird code structure, we get a
"malformed request" error message instead. Did a quick patch so that we
get the right error message page with a useful error message.

This does break the uploader for simple file uploads, but I don't think
we have to worry about that since all browsers we support should be able
to use the 'advanced' file uploader.