ADD column widths, collection specific itemdisplay
The collection home page item table now has set column widths. These
column widths aren't obeyed all that strictly due to Item Type entries
such as "Assignments/Homework" stretching it out. I didn't want to let
it word-break anywhere as that does weird splits like "Assignments/Ho"
and "mework", so we're just going to have to live with it.

The metadata shown in item display can be customized for different
collections. I've added a configuration as an example and have made
corresponding changes to the helm charts.

The webui.itemdisplay.cwsei.collections needs to be modified according
to the collections you have, the committed change is just an example.