EDIT uploaded files, review, browse list small fix
To help with the fixed column width on the browse list, added a language
attribute to the table cells and enabled auto hyphenation. This only
helps for english words that are in the browser's dictionary, but better
than nothing.

When reviewing your submission, don't escape the html tags in the
description.

Hide mentions of checksum in the uploaded files list page.