FIX blurb user not getting created
Turns out when we created the blurb user for storage, it didn't commit
the transaction, so the user is never created. This is why on prod, I
keep seeing user create and update repeatedly, as every time the
monitoring system pings the home page, it kicks off the blurb system,
which triggers the user create.