EDIT collection home table col width compat
Hacky fix: The item listing in collection home has set width columns.
Browsers don't really obey the widths for "Item Type" though cause
"Assignment/Homework" doesn't have a space in it to break on, so it is
shown in one long line. Adding a space to it helps the column obey
widths.