FIX broken pipe exceptions
Getting a lot of broken pipe exceptions from verf. No idea what causes
it, but it's always being thrown from dspace:include, so I've replaced
it with the standard jsp:include and refactored the default header to
use more JSTL. Just a shot in the dark, no clue if this is the root
cause as unable to replicate in my dev environment.