FIX xss vulnerability in search & submissions
UBC IT reported that a bug bounty hunter found that filter queries
weren't properly escaped before displaying to the user. Some research
shows that this was a regression due to heavy customization of the
search page. This suggests that other areas of heavy customization, such
as the submission & item display is probably also vulnerable.

Various text value areas now go through JSTL's escapeXml function before
being shown to the user. This fixes the original XSS demo in the bug
report.

This commit also attempts to address the other area of heavy
customization: the submissions process. The most troublesome part is
that we're using the TinyMCE editor on the frontend. Since TinyMCE
generates HTML as part of its output, we need to be able to allow some
tags through while stripping out others. I ended up using Jsoup's
whitelist cleaning feature for this. All field types that are "onebox"
and "textarea" in the submission process should now be subjected to this
cleaning. Note that this was applied broadly because it's difficult to
narrow this down to only the fields we're using TinyMCE on.

This brings up an a question: should we run escapeXml on all fields that
are not TinyMCE enabled? The only issue that prevents me from escaping
it is maintaining consistency between the submission verify step and the
actual item display. Since the verify step is very difficult to modify,
and it's currently leaving things unescaped, I've decided to leave the
item display also unescaped for now.

One exception is that the dc.subject.other field has been escaped. I
started escaping it before realizing that it's a "onebox" field and thus
should have been cleaned.